Move SDK lookup functions into separate file
These functions are a distinct group of operations that were initially
placed alongside the task planning logic, because that's what was
calling them. But there are enough of them now to justify moving them to
a separate file.
